# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DepartmentCreate": {
        "properties": {
          "code": {
            "description": "Unique Department Code, e.g. DEPT-CS",
            "maxLength": 20,
            "minLength": 2,
            "title": "Code",
            "type": "string"
          },
          "faculty_id": {
            "description": "Parent Faculty Identifier",
            "maxLength": 50,
            "minLength": 2,
            "title": "Faculty Id",
            "type": "string"
          },
          "name": {
            "description": "Department Name",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "faculty_id"
        ],
        "title": "DepartmentCreate",
        "type": "object"
      },
      "DepartmentListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/DepartmentResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "DepartmentListResponse",
        "type": "object"
      },
      "DepartmentResponse": {
        "properties": {
          "code": {
            "description": "Unique Department Code, e.g. DEPT-CS",
            "maxLength": 20,
            "minLength": 2,
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "faculty_id": {
            "description": "Parent Faculty Identifier",
            "maxLength": 50,
            "minLength": 2,
            "title": "Faculty Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "description": "Department Name",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "code",
          "name",
          "faculty_id",
          "id",
          "created_at"
        ],
        "title": "DepartmentResponse",
        "type": "object"
      },
      "DepartmentSingleResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DepartmentResponse"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "DepartmentSingleResponse",
        "type": "object"
      },
      "DepartmentUpdate": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Updated Department Code",
            "title": "Code"
          },
          "faculty_id": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Updated Faculty Identifier",
            "title": "Faculty Id"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Updated Department Name",
            "title": "Name"
          }
        },
        "title": "DepartmentUpdate",
        "type": "object"
      },
      "DepartmentValidationData": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "department_id": {
            "title": "Department Id",
            "type": "string"
          },
          "faculty_id": {
            "title": "Faculty Id",
            "type": "string"
          },
          "faculty_name": {
            "title": "Faculty Name",
            "type": "string"
          },
          "is_valid": {
            "default": true,
            "title": "Is Valid",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "department_id",
          "code",
          "name",
          "faculty_id",
          "faculty_name"
        ],
        "title": "DepartmentValidationData",
        "type": "object"
      },
      "DepartmentValidationResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DepartmentValidationData"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "DepartmentValidationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT issued by the Identity Service",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Directory microservice handling Faculties, Departments, Service Units, User Affiliations and Service Responsibilities. Versioned API under /api/v1.",
    "title": "University Directory Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/departments": {
      "get": {
        "description": "Retrieve a list of all academic departments with optional pagination and faculty filtering.",
        "operationId": "list_departments_api_v1_departments_get",
        "parameters": [
          {
            "description": "Number of items to skip",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Max items to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Max items to return",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Optional filter by parent faculty ID",
            "in": "query",
            "name": "faculty_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional filter by parent faculty ID",
              "title": "Faculty Id"
            }
          },
          {
            "description": "Case-insensitive search in code and name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search in code and name",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Departments",
        "tags": [
          "Departments"
        ]
      },
      "post": {
        "description": "Create a new academic department record associated with a valid faculty.",
        "operationId": "create_department_api_v1_departments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSingleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Department",
        "tags": [
          "Departments"
        ]
      }
    },
    "/api/v1/departments/{department_id}": {
      "delete": {
        "description": "Permanently delete a department record.",
        "operationId": "delete_department_api_v1_departments__department_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "title": "Department Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Department",
        "tags": [
          "Departments"
        ]
      },
      "get": {
        "description": "Retrieve details for a specific department by its identifier or department code.",
        "operationId": "get_department_api_v1_departments__department_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "title": "Department Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSingleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Department by ID or Code",
        "tags": [
          "Departments"
        ]
      },
      "patch": {
        "description": "Partially update details of an existing department.",
        "operationId": "patch_department_api_v1_departments__department_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "title": "Department Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSingleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Department",
        "tags": [
          "Departments"
        ]
      },
      "put": {
        "description": "Update details (name, code, faculty_id) of an existing department.",
        "operationId": "update_department_api_v1_departments__department_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "title": "Department Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentSingleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Department",
        "tags": [
          "Departments"
        ]
      }
    },
...
    "/api/v1/validation/departments/{department_id}": {
      "get": {
        "description": "Validate Department existence, identifier/code, and parent Faculty relationship for dependent services.",
        "operationId": "validate_department_api_v1_validation_departments__department_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "department_id",
            "required": true,
            "schema": {
              "title": "Department Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Department",
        "tags": [
          "Directory Validation"
        ]
      }
    },
...
  }
}